update usage of auth state
