+++ editor/hooks/index.ts
@@ -1,6 +1,5 @@
 export * from "./use-design";
 export * from "./use-async-effect";
 export * from "./use-auth-state";
-export * from "./use-figma-access-token";
 export * from "./use-target-node";
 export * from "./use-window-size";
