+++ editor/scaffolds/workspace/figma-auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./figma-auth-provider";
+export { useFigmaAuth } from "./use-figma-auth";
